pyproject.toml
[tool.poetry]
name = "tomas-bot"
version = "0.1.0"
description = "add tomas back"
authors = ["Raphael Tang <raphpb1912@gmail.com>"]
license = "ISC"
readme = "README.md"
packages = [{include = "tomas_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
Telethon = "^1.25.2"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
